requirements.txt
tensorflow
pillow
seaborn
matplotlib
opencv-python
scikit-learn
tqdm